Move lazy builder internals to Internal.Lazy.Builder
